+++ Kernel/include/queue.h
@@ -9,13 +9,13 @@
 typedef struct QueueCDT * QueueADT; // queue con roundrobin
 typedef int (*QueueElemCmpFn)(void *, void *);
 
-QueueADT createQueue(QueueElemCmpFn cmp, size_t elemSize);
+QueueADT createQueue(QueueElemCmpFn cmp, int elemSize);
 QueueADT enqueue(QueueADT queue, void * data);
 void * dequeue(QueueADT queue, void * buffer);
 void * queuePeek(QueueADT queue, void * buffer);
 void * queueRemove(QueueADT queue, void * data);
 void queueFree(QueueADT queue);
-size_t queueSize(QueueADT queue);
+int queueSize(QueueADT queue);
 QueueADT queueBeginCyclicIter(QueueADT queue);
 void * queueNextCyclicIter(QueueADT queue, void * buffer);
 int queueCyclicIterLooped(QueueADT queue);
